added epoch.go
