+++ ratelimit/keeper/abci.go
@@ -20,24 +20,3 @@ func (k Keeper) BeginBlocker(ctx sdk.Context) {
 		}
 	}
 }
-
-// Checks if it's time to start the new hour epoch
-func (k Keeper) CheckHourEpochStarting(ctx sdk.Context) (epochStarting bool, epochNumber uint64) {
-	hourEpoch := k.GetHourEpoch(ctx)
-
-	// If the block time is later than the current epoch start time + epoch duration,
-	// move onto the next epoch by incrementing the epoch number, height, and start time
-	currentEpochEndTime := hourEpoch.EpochStartTime.Add(hourEpoch.Duration)
-	shouldNextEpochStart := ctx.BlockTime().After(currentEpochEndTime)
-	if shouldNextEpochStart {
-		hourEpoch.EpochNumber++
-		hourEpoch.EpochStartTime = currentEpochEndTime
-		hourEpoch.EpochStartHeight = ctx.BlockHeight()
-
-		k.SetHourEpoch(ctx, hourEpoch)
-		return true, hourEpoch.EpochNumber
-	}
-
-	// Otherwise, indicate that a new epoch is not starting
-	return false, 0
-}
